requirements.txt
crewai
crewai_tools
pyyaml
python-dotenv
langchain_community
